queue: display current queue
We still need to connect this to (1) the number of available lines and
(2) scrollback indicator on the queue section. Imagine for instance,
that the user have over 100 songs queued (like a crazy person) therefore
we'll need to display the number available on the user terminal, and
when he hits the last one, we shall fetch more N songs from the queue,
and so on and so forth.