queue: preamble to re-rendering logic
Split between the max number of items we can display, the number of
items actually queued and only re-render if the version of the playlist
has changed (which we will use later on for fetching the data from MPD),
as when scrolling we don't need to fetch data from MPD, only update the
screen until reaches the end of the queue.

Now, seamus_frontend (horrible name), knows:

- version (the playlist version)
- queue_size (the number of items we actually have stored)

The latter, could be implemented by just hoping over
`seamus->queue`, but I think it's easier to just store the actually
number of items we have fetched.

Specially, because we have some scenarios, where: (1) you have a lot of
items queued on MPD but not enough room to display (2) have rows to
display but not enough items (3) have nothing queued at all (4) amount
of items == rows. All cases which should be handled differently,
specially on a scroll scenario.